fix perf_timer and TARGETS file

+++ monitoring/perf_timer.cc
@@ -5,8 +5,8 @@ void PerfTimer::Stop() {
     if (start_) {
       uint64_t duration = time_now() - start_;
       if (statistics_ != nullptr) {
-        if (duration > ULL(1000000000)) {
-            printf("PerfTimer::Stop duration:%d, ticker_type:%d \n", duration, ticker_type_);
+        if (duration > 1000000000ull) {
+            printf("PerfTimer::Stop duration:%lld, ticker_type:%d \n", (long long)duration, ticker_type_);
         } else {
             RecordTick(statistics_, ticker_type_, duration);
         }
